Startup REST API

+++ src/main/scala/pl/chaineural/dataStructures/package.scala
@@ -2,7 +2,7 @@ package pl.chaineural
 
 
 package object dataStructures {
-  type V = Vector[Float]
+  type V = Vector[Double]
   type M = Vector[V]
   type B = Vector[M]
 }
